% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TARGOS-package.R
\docType{package}
\name{TARGOS}
\alias{TARGOS}
\alias{TARGOS-package}
\title{A package containing Target Osteo helpers and data accessors}
\description{
This package provides data loaders for TARGET Osteosarcoma data
}

