% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TargetOsteo.R
\name{getClinical}
\alias{getClinical}
\title{Get TARGET clinical data (only 89 discovery samples) as a data frame}
\usage{
getClinical(DCC = options("DCC"))
}
\arguments{
\item{DCC}{The string location of the DCC directory}
}
\description{
Get TARGET clinical data (only 89 discovery samples) as a data frame
}
\examples{
clindf = getClinical()
head(clindf)
}

